backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.0.1
httpx==0.28.0
pydantic==2.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
websockets==13.1
aiofiles==24.1.0
requests==2.32.3
phonenumbers==8.13.50
